Modify ADA class model attribute
Update test for ada-parser
