+++ ada-parser/src/main/java/com/ucl/ADA/parser/parser/visitor/VariableDeclarationVisitor.java
@@ -6,14 +6,14 @@ import org.eclipse.jdt.core.dom.IVariableBinding;
 import org.eclipse.jdt.core.dom.VariableDeclarationFragment;
 import org.eclipse.jdt.core.dom.VariableDeclarationStatement;
 
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
+import java.util.*;
 
 @Getter
 public class VariableDeclarationVisitor extends ASTVisitor {
 
     private Map<String, String> localVariables;
+    private final String[] primitives_types = {"byte", "short", "int", "float", "double", "long", "boolean", "char"};
+    private final Set<String> PRIMITIVE_TYPES = new HashSet<String>(Arrays.asList(primitives_types));
 
 
     /**
@@ -28,7 +28,7 @@ public class VariableDeclarationVisitor extends ASTVisitor {
      * It visits the VariableDeclarationStatement node from the AST,
      * and populate a list of local variables.
      *
-     * @param node  A VariableDeclarationStatement node derived from the AST.
+     * @param node A VariableDeclarationStatement node derived from the AST.
      * @return true if it is required to visit the children node otherwise false
      */
     public boolean visit(VariableDeclarationStatement node) {
@@ -36,7 +36,10 @@ public class VariableDeclarationVisitor extends ASTVisitor {
             VariableDeclarationFragment fragment = (VariableDeclarationFragment) iter.next();
             IVariableBinding binding = fragment.resolveBinding();
             if (binding != null) {
-                localVariables.put(binding.getName(), binding.getType().getQualifiedName());
+                String type = binding.getType().getQualifiedName();
+                if ((!type.startsWith("java")) && (!PRIMITIVE_TYPES.contains(type))) {
+                    localVariables.put(binding.getName(), type);
+                }
             }
         }
         return true;
